Navigation to sides of the cube using the arrow keys is now more accurate.
I'm confused it seems to me that rotating the cube 90 degrees in any direction
should reveal at least one edge row in common between two sides of the
cube.  This doesn't seem to be happening, and I can't tell if it's
because my assumption that the edge rows should be the same, or if I
have an error in generating my cubes, or an error reading them in, or
some other kind of error.

Actually, I've just discovered that the side is correct, but the process
of rotating to the side is plagued with gimbol lock, and ends up looking
backwards (and upside down.)  Rotating the camera in the inspector, 180
degrees on the Z axis seemed to straighten the problem out, but I
haven't figured out how to do this in script yet.