Navigation to sides of the cube using the arrow keys is now more accurate.
I'm confused it seems to me that rotating the cube 90 degrees in any direction
should reveal at least one edge row in common between two sides of the
cube.  This doesn't seem to be happening, and I can't tell if it's
because my assumption that the edge rows should be the same, or if I
have an error in generating my cubes, or an error reading them in, or
some other kind of error.